requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
httpx==0.27.2
rich==13.8.1
gitpython==3.1.43
tiktoken==0.7.0
pydantic==2.9.2
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==24.1.0
anthropic==0.34.2
qdrant-client==1.12.1
